# public/assets/api.yaml
openapi: 3.0.0
info:
  title: 'My First API'
  version: '0.1'
paths:
  /failedJobs:
    get:
      tags:
        - FailedJobs
      summary: 'Find list FailedJobs'
      description: 'Returns list of FailedJobs'
      operationId: getFailedJobs
      parameters:
        -
          name: search
          in: query
          description: 'search by column defined'
          schema:
            type: object
        -
          name: order
          in: query
          description: 'order by column defined'
          schema:
            type: object
        -
          name: page
          in: query
          description: 'page to show'
          schema:
            type: int32
        -
          name: limit
          in: query
          description: 'count data display per page'
          schema:
            type: int32
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/FailedJobs' } }
                  pagination: { properties: { currentPage: { type: integer }, totalPage: { type: integer } }, type: object }
                type: object
            application/xml:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/FailedJobs' } }
                  pagination: { type: array, items: { $ref: '#/components/schemas/FailedJobs' } }
                type: object
        '404':
          description: 'FailedJobs not found'
      security:
        -
          bearer_auth: []
    post:
      tags:
        - FailedJobs
      summary: 'Add a new FailedJobs to the store'
      operationId: addFailedJobs
      requestBody:
        $ref: '#/components/requestBodies/FailedJobs'
      responses:
        '201':
          description: 'Created FailedJobs'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FailedJobs'
            application/xml:
              schema:
                $ref: '#/components/schemas/FailedJobs'
        '405':
          description: 'Invalid input'
      security:
        -
          bearer_auth: []
  '/failedJobs/{id}':
    get:
      tags:
        - FailedJobs
      summary: 'Find FailedJobs by ID'
      description: 'Returns a single FailedJobs'
      operationId: getFailedJobsById
      parameters:
        -
          name: id
          in: path
          description: 'ID of FailedJobs to return'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FailedJobs'
            application/xml:
              schema:
                $ref: '#/components/schemas/FailedJobs'
        '400':
          description: 'Invalid ID supplier'
        '404':
          description: 'FailedJobs not found'
      security:
        -
          bearer_auth: []
    put:
      tags:
        - FailedJobs
      summary: 'Update an existing FailedJobs'
      operationId: updateFailedJobs
      parameters:
        -
          name: id
          in: path
          description: 'FailedJobs id to update'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        $ref: '#/components/requestBodies/FailedJobs'
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'FailedJobs not found'
        '405':
          description: 'Validation exception'
      security:
        -
          bearer_auth: []
    delete:
      tags:
        - FailedJobs
      summary: 'Deletes a FailedJobs'
      operationId: deleteFailedJobs
      parameters:
        -
          name: id
          in: path
          description: 'FailedJobs id to delete'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'Pet not found'
      security:
        -
          bearer_auth: []
  /migrations:
    get:
      tags:
        - Migrations
      summary: 'Find list Migrations'
      description: 'Returns list of Migrations'
      operationId: getMigrations
      parameters:
        -
          name: search
          in: query
          description: 'search by column defined'
          schema:
            type: object
        -
          name: order
          in: query
          description: 'order by column defined'
          schema:
            type: object
        -
          name: page
          in: query
          description: 'page to show'
          schema:
            type: int32
        -
          name: limit
          in: query
          description: 'count data display per page'
          schema:
            type: int32
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Migrations' } }
                  pagination: { properties: { currentPage: { type: integer }, totalPage: { type: integer } }, type: object }
                type: object
            application/xml:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Migrations' } }
                  pagination: { type: array, items: { $ref: '#/components/schemas/Migrations' } }
                type: object
        '404':
          description: 'Migrations not found'
      security:
        -
          bearer_auth: []
    post:
      tags:
        - Migrations
      summary: 'Add a new Migrations to the store'
      operationId: addMigrations
      requestBody:
        $ref: '#/components/requestBodies/Migrations'
      responses:
        '201':
          description: 'Created Migrations'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Migrations'
            application/xml:
              schema:
                $ref: '#/components/schemas/Migrations'
        '405':
          description: 'Invalid input'
      security:
        -
          bearer_auth: []
  '/migrations/{id}':
    get:
      tags:
        - Migrations
      summary: 'Find Migrations by ID'
      description: 'Returns a single Migrations'
      operationId: getMigrationsById
      parameters:
        -
          name: id
          in: path
          description: 'ID of Migrations to return'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Migrations'
            application/xml:
              schema:
                $ref: '#/components/schemas/Migrations'
        '400':
          description: 'Invalid ID supplier'
        '404':
          description: 'Migrations not found'
      security:
        -
          bearer_auth: []
    put:
      tags:
        - Migrations
      summary: 'Update an existing Migrations'
      operationId: updateMigrations
      parameters:
        -
          name: id
          in: path
          description: 'Migrations id to update'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        $ref: '#/components/requestBodies/Migrations'
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'Migrations not found'
        '405':
          description: 'Validation exception'
      security:
        -
          bearer_auth: []
    delete:
      tags:
        - Migrations
      summary: 'Deletes a Migrations'
      operationId: deleteMigrations
      parameters:
        -
          name: id
          in: path
          description: 'Migrations id to delete'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'Pet not found'
      security:
        -
          bearer_auth: []
...
components:
  schemas:
    FailedJobs:
      title: FailedJobs
      description: FailedJobs
      properties:
        id:
          title: id
          description: id
          type: integer
          format: '-'
          maxLength: 20
          nullable: false
        uuid:
          title: uuid
          description: uuid
          type: string
          format: '-'
          maxLength: 255
          nullable: false
        connection:
          title: connection
          description: connection
          type: string
          format: '-'
          nullable: false
        queue:
          title: queue
          description: queue
          type: string
          format: '-'
          nullable: false
        payload:
          title: payload
          description: payload
          type: string
          format: '-'
          nullable: false
        exception:
          title: exception
          description: exception
          type: string
          format: '-'
          nullable: false
        failed_at:
          title: failed_at
          description: failed_at
          type: string
          format: '-'
          nullable: false
      type: object
    Migrations:
      title: Migrations
      description: Migrations
      properties:
        id:
          title: id
          description: id
          type: integer
          format: '-'
          maxLength: 10
          nullable: false
        migration:
          title: migration
          description: migration
          type: string
          format: '-'
          maxLength: 255
          nullable: false
        batch:
          title: batch
          description: batch
          type: integer
          format: '-'
          maxLength: 11
          nullable: false
      type: object
...
  requestBodies:
    FailedJobs:
      description: 'FailedJobs object that needs to be added'
      content:
        application/x-www-form-urlencoded:
          schema:
            $ref: '#/components/schemas/FailedJobs'
        application/xml:
          schema:
            $ref: '#/components/schemas/FailedJobs'
        application/json:
          schema:
            $ref: '#/components/schemas/FailedJobs'
    Migrations:
      description: 'Migrations object that needs to be added'
      content:
        application/x-www-form-urlencoded:
          schema:
            $ref: '#/components/schemas/Migrations'
        application/xml:
          schema:
            $ref: '#/components/schemas/Migrations'
        application/json:
          schema:
            $ref: '#/components/schemas/Migrations'
...